fix: Rename JLine backend
This commit changes the jline backend name to jline3. For consistency,
it has been moved into a `tamboui-jline3-backend` directory, similar
to the `tamboui-panama-backend` directory. The name "jline3" is in
prevision of a future JLine 4 backend.

Unfortunately, using a digit in a module name is trigerring a warning,
so the "3" is converted to "_three". Ugly but no choice.
