+++ demos/rflex-demo/src/main/java/dev/tamboui/demo/flex/RFlexDemo.java
@@ -1,5 +1,5 @@
 //DEPS dev.tamboui:tamboui-tui:LATEST
-//DEPS dev.tamboui:tamboui-jline:LATEST
+//DEPS dev.tamboui:tamboui-jline3-backend:LATEST
 /*
  * Copyright (c) 2025 TamboUI Contributors
  * SPDX-License-Identifier: MIT
