+++ tamboui-image/demos/image-demo/src/main/java/dev/tamboui/demo/ImageDemo.java
@@ -1,6 +1,6 @@
 ///usr/bin/env jbang "$0" "$@" ; exit $?
 //DEPS dev.tamboui:tamboui-image:LATEST
-//DEPS dev.tamboui:tamboui-jline:LATEST
+//DEPS dev.tamboui:tamboui-jline3-backend:LATEST
 //FILES c2023-tsuchinshan-atlas.jpg=../../../../resources/c2023-tsuchinshan-atlas.jpg
 // Prevents OSX from showing up in the terminal when running the demo
 //JAVA_OPTIONS -Dapple.awt.UIElement=true
