--- tamboui-jline/src/main/resources/META-INF/services/dev.tamboui.terminal.BackendProvider
@@ -1,1 +1,0 @@
-dev.tamboui.backend.jline.JLineBackendProvider
