+++ tamboui-jline3-backend/src/main/resources/META-INF/services/dev.tamboui.terminal.BackendProvider
@@ -1,0 +1,1 @@
+dev.tamboui.backend.jline3.JLineBackendProvider
