+++ tamboui-tfx-toolkit/demos/tfx-toolkit-demo/src/main/java/dev/tamboui/demo/TfxToolkitDemo.java
@@ -1,7 +1,7 @@
 //DEPS dev.tamboui:tamboui-tfx:LATEST
 //DEPS dev.tamboui:tamboui-tfx-toolkit:LATEST
 //DEPS dev.tamboui:tamboui-toolkit:LATEST
-//DEPS dev.tamboui:tamboui-jline:LATEST
+//DEPS dev.tamboui:tamboui-jline3-backend:LATEST
 /*
  * Copyright (c) 2025 TamboUI Contributors
  * SPDX-License-Identifier: MIT
