+++ tamboui-tfx/demos/tfx-demo/src/main/java/dev/tamboui/demo/TFXEffectsDemo.java
@@ -1,7 +1,7 @@
 ///usr/bin/env jbang "$0" "$@" ; exit $?
 //DEPS dev.tamboui:tamboui-tfx:LATEST
 //DEPS dev.tamboui:tamboui-tui:LATEST
-//DEPS dev.tamboui:tamboui-jline:LATEST
+//DEPS dev.tamboui:tamboui-jline3-backend:LATEST
 
 /*
  * Copyright (c) 2025 TamboUI Contributors
