+++ tamboui-tui/demos/helloworld-demo/src/main/java/dev/tamboui/demo/HelloWorldDemo.java
@@ -1,5 +1,5 @@
 //DEPS dev.tamboui:tamboui-widgets:LATEST
-//DEPS dev.tamboui:tamboui-jline:LATEST
+//DEPS dev.tamboui:tamboui-jline3-backend:LATEST
 
 package dev.tamboui.demo;
